Resumatch/requirements.txt
pdf2image
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
pip install pymupdf
pymupdf
